@@ -1,7 +1,3 @@
 module github.com/maretodoric/traefik-plugin-requestbodyrewrite
 
-go 1.18
-
-require (
-	github.com/traefik/plugin-sdk/v2 v2.8.4
-)
+go 1.19
