@@ -3,109 +3,103 @@ package requestbodyrewrite
 import (
     "bytes"
     "context"
+    "io"
     "io/ioutil"
     "net/http"
     "regexp"
     "strconv"
     "strings"
-
-    "github.com/traefik/plugin-sdk/v2/pkg/transport"
 )
 
-// Config holds the plugin configuration.
+// Config holds plugin configuration.
 type Config struct {
-    // A list of regex patterns and their replacements.
+    // Regex-based rewrite rules.
     Rewrites []Rewrite `json:"rewrites,omitempty"`
-    // Optional list of HTTP methods (e.g., ["POST","PUT"]) to apply rewrites to.
+    // HTTP methods to apply rewrites on, empty means all.
     Methods  []string  `json:"methods,omitempty"`
 }
 
-// Rewrite defines a single regex + replacement rule.
+// Rewrite defines a regex and its replacement text.
 type Rewrite struct {
     Regex       string `json:"regex,omitempty"`
     Replacement string `json:"replacement,omitempty"`
 }
 
-// CreateConfig returns a default configuration.
+// CreateConfig returns default Config.
 func CreateConfig() *Config {
     return &Config{}
 }
 
-// compiledRewrite holds a compiled regex for fast execution.
-type compiledRewrite struct {
-    regex       *regexp.Regexp
-    replacement string
+// compiled holds a compiled regex and replacement.
+type compiled struct {
+    re  *regexp.Regexp
+    rep string
 }
 
-// RequestBodyRewrite is the middleware instance.
+// RequestBodyRewrite plugin.
 type RequestBodyRewrite struct {
     next     http.Handler
     name     string
-    rewrites []compiledRewrite
+    rules    []compiled
     methods  map[string]struct{}
 }
 
-// New builds a new middleware instance.
+// New instantiates the plugin.
 func New(ctx context.Context, next http.Handler, config *Config, name string) (http.Handler, error) {
-    // Compile all regexes up front.
-    var rewrites []compiledRewrite
+    // Compile regex rules
+    var rules []compiled
     for _, r := range config.Rewrites {
         re, err := regexp.Compile(r.Regex)
         if err != nil {
             return nil, err
         }
-        rewrites = append(rewrites, compiledRewrite{regex: re, replacement: r.Replacement})
+        rules = append(rules, compiled{re: re, rep: r.Replacement})
     }
-
-    // Build method filter set (if provided).
+    // Build method filter
     methodsSet := make(map[string]struct{})
     for _, m := range config.Methods {
         methodsSet[strings.ToUpper(m)] = struct{}{}
     }
-
     return &RequestBodyRewrite{
-        next:     next,
-        name:     name,
-        rewrites: rewrites,
-        methods:  methodsSet,
+        next:    next,
+        name:    name,
+        rules:   rules,
+        methods: methodsSet,
     }, nil
 }
 
-// ServeHTTP reads and rewrites the request body before passing to the upstream.
+// ServeHTTP applies regex rewrites to request body.
 func (p *RequestBodyRewrite) ServeHTTP(w http.ResponseWriter, req *http.Request) {
-    // If methods filter is set and this method is not in it, skip.
+    // Skip methods not in the set
     if len(p.methods) > 0 {
         if _, ok := p.methods[req.Method]; !ok {
             p.next.ServeHTTP(w, req)
             return
         }
     }
-
     if req.Body == nil {
         p.next.ServeHTTP(w, req)
         return
     }
-
-    // Read the entire body (handles chunked, etc.).
-    bodyBytes, err := transport.ReadBody(req)
+    // Read full body
+    origBody, err := ioutil.ReadAll(req.Body)
     if err != nil {
-        // If we can't read, just forward original request.
+        // if error, reset and proceed
+        req.Body = io.NopCloser(bytes.NewReader(origBody))
         p.next.ServeHTTP(w, req)
         return
     }
-
-    // Apply each regex replacement in order.
-    bodyStr := string(bodyBytes)
-    for _, r := range p.rewrites {
-        bodyStr = r.regex.ReplaceAllString(bodyStr, r.replacement)
+    req.Body.Close()
+    // Apply rewrites
+    bodyStr := string(origBody)
+    for _, rule := range p.rules {
+        bodyStr = rule.re.ReplaceAllString(bodyStr, rule.rep)
     }
-
-    // Replace the request body and update headers.
-    newBody := []byte(bodyStr)
-    req.Body = ioutil.NopCloser(bytes.NewReader(newBody))
-    req.ContentLength = int64(len(newBody))
-    req.Header.Set("Content-Length", strconv.Itoa(len(newBody)))
-
-    // Call the next handler in the chain.
+    newBytes := []byte(bodyStr)
+    // Set new body and headers
+    req.Body = io.NopCloser(bytes.NewReader(newBytes))
+    req.ContentLength = int64(len(newBytes))
+    req.Header.Set("Content-Length", strconv.Itoa(len(newBytes)))
+    // Call next handler
     p.next.ServeHTTP(w, req)
 }
